Apply the card color through the filter instead of the mask
I updated the definition of the color filter to apply the expected
color instead of merely changing black into white. This gives sharp
results for the picture, compared to the previous blur, in Firefox.

In order to apply the same color filter to the background of the card,
I defined a mask to make the inner part of the mask transparent instead
of white (any opaque color is turned into blue by the color filter).

In this updated version, I changed the card background definition
into a mask to make the inner part of the card transparent instead
of white. I then changed the